requirements.txt
scikit-learn
streamlit_extras
tensorflow
keras==2.15.0
tensorflow